requirements.txt
pyautogen==0.2.13
requests
python-dotenv
firebase-admin
